디렉토리 구조 변경 및 세분화 작업
