+++ src/common/style.js
@@ -1,0 +1,4 @@
+export const styles = {
+  active: 'text-blue-600 font-bold',
+  inactive: 'text-gray-600',
+};
